requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
dateparser==1.2.2
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
ijson==3.2.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
yarl==1.22.0
